Add Firebase settings and update sign-in view
Added Firebase configuration settings for Android and iOS platforms to support Firebase services in the app. Created a new sign-in page for user authentication with third-party providers (Google, Facebook, and Apple) and phone number, along with the ability to switch to sign-up if an account exists already. Additionally, some color definitions in the app have been updated for further clarification.